Add scripts to start and stop a test cluster
This path adds start_cluster.py and kill_cluster.py to start and
stop a cluster respectively. Once a cluster is started, a description
of that cluster is written to a YAML file in 'running_cluster.yaml'.

When kill_cluster.py is invoked, the script reads through the YAML
file if it exists, and kills all the PIDs listed.

On every new cluster startup, there is a pre-sanity check to determine
whether a cluster is already running, and if it is, it is torn down
first before starting a new one.

Future TODOs: Add generic utils for writing YAML files instead of
the ad-hoc ways it's done in multiple places now.